requirements.txt
asgiref==3.5.0
certifi==2021.10.8
dj-database-url==0.5.0
Django==4.0.4
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
phonenumbers==8.12.47
Pillow==9.1.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
